Move VersionParsingException to API
