+++ src/main/java/net/fabricmc/loader/util/version/SemanticVersionImpl.java
@@ -17,6 +17,7 @@
 package net.fabricmc.loader.util.version;
 
 import net.fabricmc.loader.api.SemanticVersion;
+import net.fabricmc.loader.api.VersionParsingException;
 
 import java.util.Arrays;
 import java.util.Objects;
